pyproject.toml
[project]
name = "ai-essay-scorer"
version = "0.1.0"
description = ""
authors = [
    {name = "Sergey Sysoev",email = "sergey.sys@mail.ru"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "dvc (>=3.65.0,<4.0.0)",
    "kagglehub[pandas-datasets] (>=0.3.13,<0.4.0)",
    "fire (>=0.7.1,<0.8.0)",
    "torch (>=2.9.1,<3.0.0)",
    "lightning (>=2.6.0,<3.0.0)",
    "torchmetrics (>=1.8.2,<2.0.0)",
    "datasets (>=4.4.2,<5.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "transformers (>=4.57.3,<5.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)"
]

[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)",
    "ruff (>=0.14.3,<0.15.0)",
    "ipykernel (>=7.1.0,<8.0.0)",    # TODO: delete before handing in homework
    "ipywidgets (>=8.1.8,<9.0.0)",   # TODO: delete before handing in homework
]

[tool.setuptools.packages.find]
where = ["ai_essay_scorer"]

[tool.poetry]
requires-poetry = ">=2.0"
package-mode = false

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    # isort
    "I",
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # flake8-builtins
    "A",
    # flake8-return
    "RET",
    # flake8-use-pathlib
    "PTH",
    # flynt
    "FLY",
    # mccabe
    "C90",
    # NumPy-specific rules
    "NPY",
    # pep8-naming
    "N",
    # Ruff-specific rules
    "RUF",
    # Pylint - very strict, may be use when have enough time
    "PL",
    # refurb - also may be overwhelming
    "FURB",
]
ignore = [
    "PLR0913",
    "RET504",
    "N812"
    ]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 50

[tool.codespell]
skip = '*.lock,'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
